UBUNTU: SAUCE: AF_ECONET SIOCSIFADDR ioctl does not check privileges
CVE-2010-3850

In the ec_dev_ioctl function, the SIOCSIFADDR ioctl does not check
privileges, allowing an unprivileged user to assign econet addresses to
arbitrary interfaces. It is possible this is intentional, but AFAIK it
would be unusual for SIOCSIFADDR to be unprivileged, so this may be an
accident.

This patch has yet to be submitted upstream, hence marking it as SAUCE
for now.

OriginalAuthor: Phil Blundell <philb@gnu.org>

Signed-off-by: Leann Ogasawara <leann.ogasawara@canonical.com>