UBUNTU: SAUCE: sched: Prevent divide by zero when cpu_power is 0
This is a patch to fix the corner case where we're crashing with
divide_error in find_busiest_group.

I don't fully understand what the case is that causes sds.total_pwr to
be zero in find_busiest_group, but this patch guards against the
divide-by-zero bug.

I also added safe-guarding around other routines in the scheduler code
where we're dividing by power; that's more of a just-in-case and I'm
definitely open for debate on that.

BugLink: https://bugzilla.kernel.org/show_bug.cgi?id=16991
BugLink: http://bugs.launchpad.net/bugs/614853

Signed-off-by: Andrew Dickinson <whydna@whydna.net>
Signed-off-by: Stefan Bader <stefan.bader@canonical.com>
Acked-by: Andy Whitcroft <andy.whitcroft@canonical.com>
Acked-by: Tim Gardner <tim.gardner@canonical.com>
Signed-off-by: Stefan Bader <stefan.bader@canonical.com>