KVM: x86: Fix a possible backwards warp of kvmclock
Backport of commit 1d5f066e0b63271b67eac6d3752f8aa96adcbddb

Kernel time, which advances in discrete steps may progress much slower
than TSC.  As a result, when kvmclock is adjusted to a new base, the
apparent time to the guest, which runs at a much higher, nsec scaled
rate based on the current TSC, may have already been observed to have
a larger value (kernel_ns + scaled tsc) than the value to which we are
setting it (kernel_ns + 0).

We must instead compute the clock as potentially observed by the guest
for kernel_ns to make sure it does not go backwards.

Changelog:
   Mar 8: address feedback from Stefan:
	1. kernel_ns = get_kernel_ns before local_irq_restore()
	2. don't forget to vcpu->hv_clock.tsc_timestamp = tsc_timestamp;

Signed-off-by: Zachary Amsden <zamsden@redhat.com>
Signed-off-by: Marcelo Tosatti <mtosatti@redhat.com>
Signed-off-by: Serge E. Hallyn <serge.hallyn@canonical.com>
Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Tim Gardner <tim.gardner@canonical.com>