UBUNTU: (config) Disable CONFIG_NET_NS
BugLink: http://bugs.launchpad.net/bugs/720095

While network namespaces are used for containers, the implementation was
still experimental in Lucid. Even worse the cleanup of network
namespaces is very slow. This can cause severe memory issues when an
application uses them a lot (for example when cloning subprocesses).

The problem can only be avoided by backporting several upstream patches
(and without even more changes that will slow down the creation of new
net namespaces). This cannot really be justified as a stable update.

So turn the option off in Lucid. Users of containers should be pointed
to the LTS-backport kernel.

Signed-off-by: Stefan Bader <stefan.bader@canonical.com>
Acked-by: Brad Figg <brad.figg@canonical.com>
Signed-off-by: Tim Gardner <tim.gardner@canonical.com>